Move logging, tracing, and metrics into shared directory.
