+++ lsif/src/server/routes/lsif.ts
@@ -5,7 +5,7 @@ import bodyParser from 'body-parser'
 import express from 'express'
 import pTimeout from 'p-timeout'
 import uuid from 'uuid'
-import { addTags, logAndTraceCall, TracingContext } from '../../tracing'
+import { addTags, logAndTraceCall, TracingContext } from '../../shared/tracing'
 import { Backend, ReferencePaginationCursor } from '../backend/backend'
 import { DEFAULT_REFERENCES_NUM_REMOTE_DUMPS, STORAGE_ROOT } from '../settings'
 import { encodeCursor, parseCursor } from '../pagination/cursor'
