+++ lsif/src/server/server.ts
@@ -9,12 +9,12 @@ import { Backend } from './backend/backend'
 import { Connection } from 'typeorm'
 import { createDumpRouter } from './routes/dumps'
 import { createJobRouter } from './routes/jobs'
-import { createLogger } from '../logging'
+import { createLogger } from '../shared/logging'
 import { createLsifRouter } from './routes/lsif'
 import { createMetaRouter } from './routes/meta'
 import { createPostgresConnection } from '../connection'
 import { createQueue, ensureOnlyRepeatableJob } from '../queue'
-import { createTracer } from '../tracing'
+import { createTracer } from '../shared/tracing'
 import { dbFilename, dbFilenameOld, ensureDirectory } from '../util'
 import { default as tracingMiddleware } from 'express-opentracing'
 import { defineRedisCommands } from './redis/redis'
