+++ lsif/src/shared/logging.ts
