+++ lsif/src/shared/metrics.ts
