+++ lsif/src/worker/worker.ts
@@ -1,17 +1,17 @@
 import * as constants from '../constants'
 import * as path from 'path'
 import promClient from 'prom-client'
-import { addTags, createTracer, logAndTraceCall, TracingContext } from '../tracing'
+import { addTags, createTracer, logAndTraceCall, TracingContext } from '../shared/tracing'
 import { createCleanFailedJobsProcessor } from './processors/clean-failed-jobs'
 import { createCleanOldJobsProcessor } from './processors/clean-old-jobs'
 import { createConvertJobProcessor } from './processors/convert'
-import { createLogger } from '../logging'
+import { createLogger } from '../shared/logging'
 import { createPostgresConnection } from '../connection'
 import { createQueue } from '../queue'
 import { createUpdateTipsJobProcessor } from './processors/update-tips'
 import { ensureDirectory } from '../util'
 import { followsFrom, FORMAT_TEXT_MAP, Span, Tracer } from 'opentracing'
-import { instrument } from '../metrics'
+import { instrument } from '../shared/metrics'
 import { Job } from 'bull'
 import { jobDurationErrorsCounter, jobDurationHistogram } from '../worker.metrics'
 import { Logger } from 'winston'
